We are Learn How to  Perform Fibonacci_Series  In java

+++ src/Fibonacci_Series.java
@@ -6,7 +6,7 @@ public class Fibonacci_Series {
         System.out.println("Welcome to Fibonacci Series\n________________________");
         System.out.print("Enter Your Number: ");
         int num = input.nextInt();
-        int fib = Fib_Num(num);
+        long fib = Fib_Num(num);
         System.out.println("Fibonacci Series up to " + num + " terms:");
 
     }
